+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 IMPROVEMENTS
 * [\#53](https://github.com/binance-chain/go-sdk/pull/53) [SOURCE] change the default source into 0
 * [\#56](https://github.com/binance-chain/go-sdk/pull/56) [RPC] add reconnect strategy when timeout to receive response
+
+## BREAKING
+* [\#57](https://github.com/binance-chain/go-sdk/pull/57) [API] add query option to getTokens api